Modified Tree for new implemenation without eventlist, and to take in new csv format
Editted the Tree implementation to take in new csv format, as well as
allow tree to take over the role of event list by storing all events
directly on the tree.

Renamed some of the functions on the tree to make the function name more
descriptive

Changed the implementation of data storage on the nodes to use an
Ordered Dictionary using the hash of the event as a key, and the entire
event as a value.